evaluation/requirements.txt
# Evaluation Framework Dependencies

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# API clients for LLM judge
openai>=1.0.0  # For GPT-4 judge (paid)
anthropic>=0.18.0  # For Claude judge (paid)
google-genai>=1.0.0  # For Gemini judge (FREE - recommended for testing)

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Colored terminal output

# Optional: For web dashboard
# flask>=3.0.0
# plotly>=5.18.0
